Cleaned up openai external api call functions, put them in a new file called openai-service. Also worked on making my server file as tip top as possible.
